+++ fitsio/test.py
@@ -1348,7 +1348,7 @@ class TestReadWrite(unittest.TestCase):
         Insert a new column
         """
 
-        fname=tempfile.mktemp(prefix='fitsio-TableDeleteRowRange-',suffix='.fits')
+        fname=tempfile.mktemp(prefix='fitsio-TableDeleteRows-',suffix='.fits')
         try:
             with fitsio.FITS(fname,'rw',clobber=True) as fits:
                 fits.write_table(self.data)
@@ -1362,6 +1362,96 @@ class TestReadWrite(unittest.TestCase):
 
             compare_data = self.data[ [0,2] ]
             self.compare_rec(compare_data, d, "delete rows")
+
+
+        finally:
+            if os.path.exists(fname):
+                os.remove(fname)
+
+    def testTableResize(self):
+        """
+        Insert a new column
+        """
+
+        fname=tempfile.mktemp(prefix='fitsio-TableResize-',suffix='.fits')
+        try:
+
+            #
+            # shrink from back
+            #
+            with fitsio.FITS(fname,'rw',clobber=True) as fits:
+                fits.write_table(self.data)
+
+            nrows = 2
+            with fitsio.FITS(fname,'rw') as fits:
+                fits[1].resize(nrows)
+
+            with fitsio.FITS(fname) as fits:
+                d = fits[1].read()
+
+            compare_data = self.data[0:nrows]
+            self.compare_rec(compare_data, d, "shrink from back")
+
+
+            #
+            # shrink from front
+            #
+            with fitsio.FITS(fname,'rw',clobber=True) as fits:
+                fits.write_table(self.data)
+
+            with fitsio.FITS(fname,'rw') as fits:
+                fits[1].resize(nrows, front=True)
+
+            with fitsio.FITS(fname) as fits:
+                d = fits[1].read()
+
+            compare_data = self.data[nrows-self.data.size:]
+            self.compare_rec(compare_data, d, "shrink from front")
+
+
+            # These don't get zerod
+
+            nrows = 10
+            add_data = numpy.zeros(nrows-self.data.size,dtype=self.data.dtype)
+            add_data['i1scalar'] = -128
+            add_data['i1vec'] = -128
+            add_data['i1arr'] = -128
+            add_data['u2scalar'] = 32768
+            add_data['u2vec'] = 32768
+            add_data['u2arr'] = 32768
+            add_data['u4scalar'] = 2147483648
+            add_data['u4vec'] = 2147483648
+            add_data['u4arr'] = 2147483648
+
+
+            #
+            # expand at the back
+            #
+            with fitsio.FITS(fname,'rw',clobber=True) as fits:
+                fits.write_table(self.data)
+            with fitsio.FITS(fname,'rw') as fits:
+                fits[1].resize(nrows)
+
+            with fitsio.FITS(fname) as fits:
+                d = fits[1].read()
+
+            compare_data = numpy.hstack( (self.data, add_data) )
+            self.compare_rec(compare_data, d, "expand at the back")
+
+            #
+            # expand at the front
+            #
+            with fitsio.FITS(fname,'rw',clobber=True) as fits:
+                fits.write_table(self.data)
+            with fitsio.FITS(fname,'rw') as fits:
+                fits[1].resize(nrows, front=True)
+
+            with fitsio.FITS(fname) as fits:
+                d = fits[1].read()
+
+            compare_data = numpy.hstack( (add_data, self.data) )
+            # These don't get zerod
+            self.compare_rec(compare_data, d, "expand at the front")
 
 
         finally: